# docs/api/api_docs/openapi.json
{
  "openapi": "3.0.2",
  "info": {
    "title": "AI Interviewer Agent API",
    "description": "API for AI interview preparation services",
    "version": "1.0.0"
  },
  "paths": {
    "/api/interview/start": {
      "post": {
        "summary": "Start Interview",
        "description": "Start a new interview session.\n\nArgs:\n    config: Interview configuration\n    db: Database session\n    \nReturns:\n    Information about the created session",
        "operationId": "start_interview_api_interview_start_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InterviewConfig"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionInfo"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/interview/send": {
      "post": {
        "summary": "Send Message",
        "description": "Send a message to the interview agent.\n\nArgs:\n    message: The user's message\n    background_tasks: FastAPI background tasks\n    db: Database session\n    orchestrator: Orchestrator instance\n    \nReturns:\n    Agent response",
        "operationId": "send_message_api_interview_send_post",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            },
            "name": "session_id",
            "in": "query"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserMessage"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/interview/end": {
      "post": {
        "summary": "End Interview",
        "description": "End an interview session.\n\nArgs:\n    message: The user's message (session_id required)\n    background_tasks: FastAPI background tasks\n    db: Database session\n    orchestrator: Orchestrator instance\n    \nReturns:\n    End confirmation and summary",
        "operationId": "end_interview_api_interview_end_post",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            },
            "name": "session_id",
            "in": "query"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserMessage"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response End Interview Api Interview End Post",
                  "type": "object"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/interview/sessions": {
      "get": {
        "summary": "List Sessions",
        "description": "List active interview sessions.\n\nArgs:\n    user_id: Filter by user ID\n    \nReturns:\n    List of active sessions",
        "operationId": "list_sessions_api_interview_sessions_get",
        "parameters": [
          {
            "required": false,
            "schema": {
              "title": "User Id",
              "type": "string"
            },
            "name": "user_id",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response List Sessions Api Interview Sessions Get",
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SessionInfo"
                  }
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/interview/info": {
      "get": {
        "summary": "Get Session Info",
        "description": "Get information about an interview session.\n\nArgs:\n    session_id: The session identifier\n    \nReturns:\n    Session information",
        "operationId": "get_session_info_api_interview_info_get",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            },
            "name": "session_id",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionInfo"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/interview/history": {
      "get": {
        "summary": "Get Conversation History",
        "description": "Get the conversation history for a session.\n\nArgs:\n    session_id: The session identifier\n    orchestrator: Orchestrator instance\n    \nReturns:\n    List of conversation messages",
        "operationId": "get_conversation_history_api_interview_history_get",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            },
            "name": "session_id",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Get Conversation History Api Interview History Get",
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/interview/skill-assessment": {
      "post": {
        "summary": "Start Skill Assessment",
        "description": "Start a skill assessment session or switch existing session to skill assessment mode.\n\nArgs:\n    message: User message with session ID\n    background_tasks: FastAPI background tasks\n    db: Database session\n    \nReturns:\n    Skill assessment initialization confirmation",
        "operationId": "start_skill_assessment_api_interview_skill_assessment_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserMessage"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Start Skill Assessment Api Interview Skill Assessment Post",
                  "type": "object"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/interview/skills": {
      "get": {
        "summary": "Get Skills",
        "description": "Get skills assessed in an interview session.\n\nArgs:\n    session_id: The session identifier\n    db: Database session\n    \nReturns:\n    List of skill assessments",
        "operationId": "get_skills_api_interview_skills_get",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            },
            "name": "session_id",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Get Skills Api Interview Skills Get",
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SkillAssessment"
                  }
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "title": "HTTPValidationError",
        "type": "object",
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            }
          }
        }
      },
      "InterviewConfig": {
        "title": "InterviewConfig",
        "required": [
          "job_role"
        ],
        "type": "object",
        "properties": {
          "job_role": {
            "title": "Job Role",
            "type": "string",
            "description": "Target job role for the interview"
          },
          "job_description": {
            "title": "Job Description",
            "type": "string",
            "description": "Detailed job description",
            "default": ""
          },
          "interview_style": {
            "title": "Interview Style",
            "type": "string",
            "description": "Style of interview (FORMAL, CASUAL, AGGRESSIVE, TECHNICAL)",
            "default": "FORMAL"
          },
          "mode": {
            "title": "Mode",
            "type": "string",
            "description": "Operating mode for the orchestrator",
            "default": "interview"
          },
          "user_id": {
            "title": "User Id",
            "type": "string",
            "description": "User identifier"
          }
        },
        "description": "Model for configuring a new interview session."
      },
...
    }
  }
}